Updated Planned Stuff
So stuff that is set up:
-Posting a Ride

[5:28]
-Posting a Ride
Request

[5:30]
--Grabbing all the Riders and Joining a Ride

[5:30]

What doesn't work:
-view all the rides that you posted yourself
(I will
finish that when I wake up)

[5:32]
Also my suggestion for tomorrow is
if we have mics jump on to some sort of voice thing to get it
working.

[5:32]
Or meet in person up to you guys.

[5:36]
Also
there is a weird bug where it says there is no member for a model
despite the model being there.